SizeMappingTests shouldn't use ESSingleNodeTestCase
